+++ sickbeard/providers/limetorrents.py
@@ -67,6 +67,8 @@ class LimeTorrentsProvider(generic.TorrentProvider):
                     else self.urls['search'] % (urllib.quote_plus(search_string))
 
                 html = self.get_url(search_url)
+                if self.should_skip():
+                    return results
 
                 cnt = len(items[mode])
                 try: