+++ sickbeard/providers/pretome.py
@@ -16,7 +16,6 @@
 # along with SickGear.  If not, see <http://www.gnu.org/licenses/>.
 
 from . import generic
-from sickbeard.rssfeeds import RSSFeeds
 from lib.unidecode import unidecode
 
 
@@ -52,7 +51,7 @@ class PreToMeProvider(generic.TorrentProvider):
                 search_string = isinstance(search_string, unicode) and unidecode(search_string) or search_string
                 search_url = url + (self.urls['search'] % search_string, '')['Cache' == mode]
 
-                xml_data = RSSFeeds(self).get_feed(search_url)
+                xml_data = self.cache.get_rss(search_url)
 
                 cnt = len(items[mode])
                 if xml_data and 'entries' in xml_data: