+++ sickbeard/providers/torrenting.py
@@ -69,6 +69,8 @@ class TorrentingProvider(generic.TorrentProvider):
                 search_url = self.urls['search'] % (self._categories_string(), search_string)
 
                 html = self.get_url(search_url)
+                if self.should_skip():
+                    return results
 
                 cnt = len(items[mode])
                 try: