+++ sickbeard/providers/torrentleech.py
@@ -62,6 +62,8 @@ class TorrentLeechProvider(generic.TorrentProvider):
                     'query': isinstance(search_string, unicode) and unidecode(search_string) or search_string}
 
                 html = self.get_url(search_url)
+                if self.should_skip():
+                    return results
 
                 cnt = len(items[mode])
                 try: